Implement one round of rock paper scissors
Created index.html and javascript.js. Implemented the logic that
prompts the user for their move. The logic checks if the user input is
valid and makes it lower case. The script then evaluates the choice of
the user and the random choice of the computer. The playRound function
can now select the winner and increment the scores.